+++ meteoio/plugins/GrassIO.cc
@@ -72,7 +72,7 @@ void GrassIO::read2DGrid(Grid2DObject& grid_out, const std::string& filename)
 {
 
 	int _nx, _ny;
-	unsigned int ncols, nrows;
+	size_t ncols, nrows;
 	double north, east, south, west;
 	double tmp_val, xllcorner, yllcorner, cellsize;
 	vector<string> tmpvec;
@@ -117,8 +117,8 @@ void GrassIO::read2DGrid(Grid2DObject& grid_out, const std::string& filename)
 		if((_nx<0) || (_ny<0)) {
 			throw IOException("Number of rows or columns in 2D Grid read as \"nodata\", in file: " + filename, AT);
 		}
-		ncols = (unsigned int)_nx;
-		nrows = (unsigned int)_ny;
+		ncols = (size_t)_nx;
+		nrows = (size_t)_ny;
 		xllcorner = west;
 		yllcorner = south;
 		cellsize = (east - west) / (double)ncols;
@@ -131,14 +131,14 @@ void GrassIO::read2DGrid(Grid2DObject& grid_out, const std::string& filename)
 		grid_out.set(ncols, nrows, cellsize, coordinate);
 
 		//Read one line after the other and parse values into Grid2DObject
-		for (unsigned int kk=nrows-1; (kk < nrows); kk--) {
+		for (size_t kk=nrows-1; (kk < nrows); kk--) {
 			getline(fin, line, eoln); //read complete line
 
 			if (IOUtils::readLineToVec(line, tmpvec) != ncols) {
 				throw InvalidFormatException("Premature End " + filename, AT);
 			}
 
-			for (unsigned int ll=0; ll < ncols; ll++){
+			for (size_t ll=0; ll < ncols; ll++){
 				if (tmpvec[ll] == "*"){
 					tmp_val = plugin_nodata;
 				} else {
